refactor: decompor TransactionLive.Index e consolidar backlog
Extrai os componentes da view de transações para LiveComponents menores e arquivos HEEx dedicados, melhorando a manutenibilidade. Além disso, consolida todos os TODOs espalhados pelo repositório num único arquivo BACKLOG.md, junto com as diretrizes do projeto.
