+++ lib/cash_lens/parsers/ingestor.ex
@@ -4,6 +4,8 @@ defmodule CashLens.Parsers.Ingestor do
   """
   alias CashLens.Parsers.CSVParser
   alias CashLens.Parsers.PDFParser
+  alias CashLens.Transactions
+  alias CashLens.Accounts
 
   @doc """
   Parses the content based on the provided parser_type.
@@ -21,5 +23,70 @@ defmodule CashLens.Parsers.Ingestor do
       _ ->
         {:error, "Extrator não configurado ou não suportado para esta conta."}
     end
+  end
+
+  @doc """
+  Reads a file, converts encoding/extracts text, parses and saves the transactions.
+  Returns `{:ok, count}` or `{:error, reason}`.
+  """
+  def import_file(account, file_path) do
+    content = File.read!(file_path)
+
+    content =
+      if String.ends_with?(file_path, ".pdf") or account.parser_type == "sem_parar_pdf" do
+        case System.cmd("pdftotext", ["-layout", file_path, "-"]) do
+          {text, 0} -> text
+          _ -> content
+        end
+      else
+        if String.valid?(content),
+          do: content,
+          else: :unicode.characters_to_binary(content, :latin1, :utf8)
+      end
+
+    case parse(content, account.parser_type) do
+      {:error, reason} ->
+        {:error, reason}
+
+      transactions_data ->
+        periods = process_transactions_data(transactions_data, account.id)
+
+        periods
+        |> MapSet.to_list()
+        |> Enum.each(fn {acc_id, month, year} ->
+          CashLens.Accounting.calculate_monthly_balance(acc_id, year, month)
+        end)
+
+        {:ok, length(transactions_data)}
+    end
+  end
+
+  defp process_transactions_data(transactions_data, account_id) do
+    Enum.reduce(transactions_data, MapSet.new(), fn data, acc ->
+      data
+      |> Map.put(:account_id, account_id)
+      |> CashLens.Transactions.AutoCategorizer.categorize()
+      |> Transactions.create_transaction()
+
+      acc = MapSet.put(acc, {account_id, data.date.month, data.date.year})
+      description = String.upcase(data.description || "")
+
+      cond do
+        String.contains?(description, "BB MM OURO") ->
+          case Accounts.get_account_by_name("BB MM Ouro") do
+            nil -> acc
+            a -> MapSet.put(acc, {a.id, data.date.month, data.date.year})
+          end
+
+        String.contains?(description, ["BB RENDE FÁCIL", "BB RENDE FACIL"]) ->
+          case Accounts.get_account_by_name("BB Rende Fácil") do
+            nil -> acc
+            a -> MapSet.put(acc, {a.id, data.date.month, data.date.year})
+          end
+
+        true ->
+          acc
+      end
+    end)
   end
 end
